change: moved `service` and `ssl` source file to `http` folder.
